+++ src/Http/Middleware/NegotiatesContent.php
@@ -22,11 +22,12 @@ use Neomerx\JsonApi\Contracts\Codec\CodecMatcherInterface;
 use Neomerx\JsonApi\Contracts\Http\HttpFactoryInterface;
 use Neomerx\JsonApi\Exceptions\JsonApiException;
 use Psr\Http\Message\ServerRequestInterface;
+use function CloudCreativity\JsonApi\http_contains_body;
 
 /**
  * Trait NegotiatesContent
  *
- * @package CloudCreativity\JsonApi\Http\Middleware
+ * @package CloudCreativity\JsonApi
  */
 trait NegotiatesContent
 {
@@ -48,6 +49,6 @@ trait NegotiatesContent
         $parser = $httpFactory->createHeaderParametersParser();
         $checker = $httpFactory->createHeadersChecker($codecMatcher);
 
-        $checker->checkHeaders($parser->parse($request));
+        $checker->checkHeaders($parser->parse($request, http_contains_body($request)));
     }
 }